requirements.txt
basicsr==1.4.2
minio==7.2.7
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.10.0.82
paddleocr==2.8.0
paddleocr.egg==info
pip==24.0
psutil==5.9.8
PyYAML==6.0.1
setuptools==69.5.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
